refactor: use buffer lib for deprecated methods; change error messages
