+++ package.json
@@ -67,6 +67,7 @@
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
     "add-js-extension": "^1.0.4",
+    "buffer": "^6.0.3",
     "c8": "^7.13.0",
     "esbuild": "^0.19.8",
     "eslint": "^8.55.0",
