+++ src/processors/symmetricEncrypt.ts
